backend/requirements.txt
django-cors-headers==3.7.0
Django>=3.2,<4.0
social-auth-app-django>=5.4,<6.0
requests>=2.9.1
dj-rest-auth==5.0.2
django-allauth==0.47.0
djangorestframework-simplejwt==5.0.0
django-oauth-toolkit==1.5.0
python-dotenv==0.19.0
